Refactor zha.core.helpers.
