+++ homeassistant/components/zha/config_flow.py
@@ -1,10 +1,21 @@
 """Config flow for ZHA."""
+import asyncio
 from collections import OrderedDict
 import os
 
+import bellows.ezsp
+import bellows.zigbee
+
 import voluptuous as vol
+import zigpy_deconz.api
+import zigpy_deconz.zigbee
+import zigpy_xbee.api
+import zigpy_xbee.zigbee
+import zigpy_zigate.api
+import zigpy_zigate.zigbee
 
 from homeassistant import config_entries
+from .core.const import DEFAULT_BAUDRATE
 
 from .core.const import (
     CONF_RADIO_TYPE,
@@ -13,7 +24,6 @@ from .core.const import (
     DOMAIN,
     RadioType,
 )
-from .core.helpers import check_zigpy_connection
 
 
 @config_entries.HANDLERS.register(DOMAIN)
@@ -57,3 +67,27 @@ class ZhaFlowHandler(config_entries.ConfigFlow):
         return self.async_create_entry(
             title=import_info[CONF_USB_PATH], data=import_info
         )
+
+
+async def check_zigpy_connection(usb_path, radio_type, database_path):
+    """Test zigpy radio connection."""
+    if radio_type == RadioType.ezsp.name:
+        radio = bellows.ezsp.EZSP()
+        ControllerApplication = bellows.zigbee.application.ControllerApplication
+    elif radio_type == RadioType.xbee.name:
+        radio = zigpy_xbee.api.XBee()
+        ControllerApplication = zigpy_xbee.zigbee.application.ControllerApplication
+    elif radio_type == RadioType.deconz.name:
+        radio = zigpy_deconz.api.Deconz()
+        ControllerApplication = zigpy_deconz.zigbee.application.ControllerApplication
+    elif radio_type == RadioType.zigate.name:
+        radio = zigpy_zigate.api.ZiGate()
+        ControllerApplication = zigpy_zigate.zigbee.application.ControllerApplication
+    try:
+        await radio.connect(usb_path, DEFAULT_BAUDRATE)
+        controller = ControllerApplication(radio, database_path)
+        await asyncio.wait_for(controller.startup(auto_form=True), timeout=30)
+        await controller.shutdown()
+    except Exception:  # pylint: disable=broad-except
+        return False
+    return True
